DocSum/DocumentSummaryFlaskBackend/requirements.txt
Flask==2.0.3
Flask-Cors==3.0.10
nltk==3.6.7
pandas==1.3.5
numpy==1.21.5
sentence-transformers==2.1.0
PyPDF2==1.26.0
google-generativeai==0.1.0